zhaohui

+++ src/org/ldxx/bean/StatisticsVo.java
@@ -17,48 +17,47 @@ package org.ldxx.bean;
 */
 public class StatisticsVo {
 
-	private float mbContractMoney;//目标合同金额
-	private float mbIncome;//目标收入
-	private float mbReceipt;//目标收款
-	private float contractAll;//实际总合同金额
-	private float allInCome;//实际总收入
-	private float allReceipt;//实际总收款
-	public float getMbContractMoney() {
+	private double mbContractMoney;//目标合同金额
+	private double mbIncome;//目标收入
+	private double mbReceipt;//目标收款
+	private double contractAll;//实际总合同金额
+	private double allInCome;//实际总收入
+	private double allReceipt;//实际总收款
+	public double getMbContractMoney() {
 		return mbContractMoney;
 	}
-	public void setMbContractMoney(float mbContractMoney) {
+	public void setMbContractMoney(double mbContractMoney) {
 		this.mbContractMoney = mbContractMoney;
 	}
-	public float getMbIncome() {
+	public double getMbIncome() {
 		return mbIncome;
 	}
-	public void setMbIncome(float mbIncome) {
+	public void setMbIncome(double mbIncome) {
 		this.mbIncome = mbIncome;
 	}
-	public float getMbReceipt() {
+	public double getMbReceipt() {
 		return mbReceipt;
 	}
-	public void setMbReceipt(float mbReceipt) {
+	public void setMbReceipt(double mbReceipt) {
 		this.mbReceipt = mbReceipt;
 	}
-	public float getContractAll() {
+	public double getContractAll() {
 		return contractAll;
 	}
-	public void setContractAll(float contractAll) {
+	public void setContractAll(double contractAll) {
 		this.contractAll = contractAll;
 	}
-	public float getAllInCome() {
+	public double getAllInCome() {
 		return allInCome;
 	}
-	public void setAllInCome(float allInCome) {
+	public void setAllInCome(double allInCome) {
 		this.allInCome = allInCome;
 	}
-	public float getAllReceipt() {
+	public double getAllReceipt() {
 		return allReceipt;
 	}
-	public void setAllReceipt(float allReceipt) {
+	public void setAllReceipt(double allReceipt) {
 		this.allReceipt = allReceipt;
 	}
-	
 	
 }
